Further work on animations
Replaced many of the animations used with the child models, so that now they are no longer stretched.

Also replaced two of the listening animations due to them being not suitable for what we are trying to achieve. And, replaced them with nodding animations.

Worked on refining the transitions between animations but there is still a bit of work to be done.