Implemented prompted conversation
Implemented a version of the current use of GPT-3 that loosens the reigns on what it can say. Where rather than being asked to generate a variation of a given text, it is being asked to generate its own responses based on the conversation thus far. Taking into account its own personal knowledge of the situation which is stored in a new string variable.

Also, made it so that you can assign characters to not have what they say added to the transcription of the conversation. Since, for our purposes we do not want what the teacher figure is saying to influence what the other NPCs are saying. Thus, we wanted her ignored.

We have kept things constrained for the user and the teacher figure in terms of what can be said. Since we want our implementation to be comparable, and for their to be a planned end. Hence why the user can still only choose one of three options, and why the teacher will always at the very most (when it comes to GPT) use text variation.