Implemented check for prompted conversation
Essentially, during a prompted conversation during transition points from one person to another we found that rather than just writing a response for themselves. A character might follow the pattern of conversation up until that point and try to 'anticipate' how the conversation may go.
Creating a response that could include multiple back and forth between the other character and the user. Often for the sake of putting themselves in a better position (which was rather interesting).

Nevertheless, this was not something we wanted. So we decided to switch a key character involved in formatting the text to something that would be extremely unlikely to ever be used for any other purpose other than those that we have placed (the '|').
We did this so that we can easily search for this in a split and if this resulted in an array above a size of one then we knew that it has included a response from at least one other character.

While this is not an ideal situation, since it is still generating it as part of the response (thus using tokens), this seems like the best and most consistent manner of approaching this.