Replaced boy and cleaned up animations
Replaced the child with a shrunk adult which seems to work well.

Made some adjustments to the way the idle animations are handled in the controller so that they loop in a nicer manner.

Also made use of the RigidBody component to help keep the characters in place. Reducing some of the rather distracting movement that is present.