Replaced some of two idle animations and one talking
These animations often moved the characters in a rather undesirable way that did not mesh well with the other animations. Or, by using them it led to things like clipping through objects.

For one, both.

So we decided it would be best to try and replace these animations with something that be more natural to look at and more in-line with the other animations that are already being used.